refactor(pack): rename 'pack.lsp' to 'pack._lsp'
